Move service handlers to dedicated folders
