+++ plugins/bulk-import-backend/src/service/handlers/ping/ping.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2024 The Janus IDP Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { LoggerService } from '@backstage/backend-plugin-api';
+
+import type { Paths } from '../../../generated/openapi';
+import type { HandlerResponse } from '../handlers';
+
+export async function ping(
+  logger: LoggerService,
+): Promise<HandlerResponse<Paths.Ping.Responses.$200>> {
+  logger.debug('PONG!');
+  return {
+    statusCode: 200,
+    responseBody: { status: 'ok' },
+  };
+}
