+++ plugins/bulk-import-backend/src/service/handlers/repository/index.ts
@@ -14,17 +14,4 @@
  * limitations under the License.
  */
 
-import type { LoggerService } from '@backstage/backend-plugin-api';
-
-import type { Paths } from '../../generated/openapi.d';
-import type { HandlerResponse } from './handlers';
-
-export async function ping(
-  logger: LoggerService,
-): Promise<HandlerResponse<Paths.Ping.Responses.$200>> {
-  logger.debug('PONG!');
-  return {
-    statusCode: 200,
-    responseBody: { status: 'ok' },
-  };
-}
+export * from './repositories';
